move textures to separate lua module
